wip: useCreateMap hook
